requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
pydantic==2.12.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.31.0
PyJWT==2.8.0

